pyproject.toml
[tool.poetry]
name = "python-clean-architecture"
version = "0.1.0"
description = "This structure will help in the project building using python and fastapi to isolate the dependencies and have a cleaner code"
authors = ["Sebastian Ituarte <sebastian.ituarte@yapo.cl>"]
readme = "README.md"
packages = [{ include = "python-clean-architecture" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.93.0"
uvicorn = { extras = ["standard"], version = "^0.19.0" }
sqlalchemy = "^1.4.42"
alembic = "^1.8.1"
email-validator = "^1.3.0"
psycopg2-binary = "^2.9.5"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.2"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
dependency-injector = "^4.41.0"
aioredis = "^2.0.1"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.10.3"
flake8 = "^5.0.4"

[tool.taskipy.tasks]
dev = "uvicorn app.main:app --reload"
prod = "uvicorn app.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"